size box

+++ inc/js/profile-edit.js
@@ -132,7 +132,9 @@ jQuery(document).ready(function($) {
     });
 
     $(document).on('change','#fileUpload-file', function() {
-        jcrop_api.destroy();
+        if (jcrop_api) {
+            jcrop_api.destroy();
+        }
         var fileData = $('#fileUpload-fileData');
         var filePreview = $('#fileUpload-preview');
         var previewFields = $('.fileUpload-preview');
